refactor: adopt src/ layout and cvbench package namespace
