+++ .dockerignore
@@ -1,20 +1,10 @@
 # Exclude all.
 *
 
-# Except core framework
-!core/
-!core/**/
-!core/**/*.py
-
-# Except CLI
-!cli/
-!cli/**/
-!cli/**/*.py
-
-# Except augmentations
-!augmentations/
-!augmentations/**/
-!augmentations/**/*.py
+# Except source package
+!src/
+!src/**/
+!src/**/*.py
 
 # Except tests
 !tests/
