--- augmentations/__init__.py
@@ -1,6 +1,0 @@
-from augmentations.tone     import *    # noqa: F401, F403
-from augmentations.noise    import *    # noqa: F401, F403
-from augmentations.lines    import *    # noqa: F401, F403
-from augmentations.edges    import *    # noqa: F401, F403
-from augmentations.blur     import *    # noqa: F401, F403
-from augmentations.profiles import *    # noqa: F401, F403
