+++ src/cvbench/augmentations/edges.py
