+++ src/cvbench/augmentations/lines.py
