+++ src/cvbench/augmentations/noise.py
