+++ src/cvbench/augmentations/profiles.py
