+++ src/cvbench/augmentations/tone.py
