+++ src/cvbench/cli/__init__.py
