+++ src/cvbench/cli/augmentations.py
