+++ src/cvbench/cli/generate.py
