+++ src/cvbench/cli/predict.py
