+++ src/cvbench/core/evaluator.py
